refactor(routes): merge file upload routes under /storage prefix
